Bunch of changes to make final build work.

+++ WoWPetBattler/PetHelper.h
@@ -13,7 +13,7 @@ public:
 	Q_INVOKABLE static void CheckHealing(PetStage*, quint8, quint8, quint16, bool);
 	static void CheckRacialsProc(PetStage*, quint8);
 	static void CheckRacials(Pet*);
-	Q_INVOKABLE static void CheckAuraPower(PetStage*, PetAura*, quint8, quint8);
+	Q_INVOKABLE static void CheckAuraPower(PetStage*, PetAura*, quint8, quint16);
 	Q_INVOKABLE static void CheckCleansingRain(PetStage*, quint8, quint8, quint16, qint8, bool, quint16);
 	Q_INVOKABLE static float CheckWeatherBonus(PetStage*, PetType::Type);
 };
